whisper-service/requirements.txt
flask==3.0.0
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2
numpy==1.24.3
ffmpeg-python==0.2.0
gunicorn==21.2.0
Werkzeug==3.0.1